- add support for refresh tokens - rename JWTUtil.java to TokenServiceImpl.Java   - interactions with refresh tokens now also handled by TokenService - access JWTs now granted in secured HTTP-Only cookies rather than returned as string to client   - refresh token also stored in HTTP-Only cookie - refactor JWTAuthenticationFilter.java to integrate with TokenService implementations - reclassified *Dao classes as *Repository and changed path to reflect this - adjust order of & logic of SecurityFilterChains to properly restrict api endpoints - fix issue with registration caused by new use of provider info - fix issues with exception handling - fix issue where queries for items were producing duplicate image entries - add signout endpoint - refactor AuthResponse.java to deliver entire User object as payload - refactor ImageSerializer.java to deliver more info about Image - update hash/ isEqual methods in several classes to prevent bugs - fix issue where deleting an Image caused a loop that emitted an error - add endpoint for uploading new images to existing items - add method security to findItemsByUser method in ItemServiceImpl.java - refactor signout endpoint to not rely on Authentication object - add bean for mail sender - add service for sending email - add mechanisms & DB columns for email verification - minor bug fixes/ code cleanup - alter /verifyRegistration endpoint to return auth credentials - alter VerificationToken to be a 6-char hex value

+++ src/main/java/com/qronicle/controller/ItemController.java
@@ -4,33 +4,39 @@ import com.qronicle.entity.Image;
 import com.qronicle.entity.Item;
 import com.qronicle.entity.Tag;
 import com.qronicle.entity.User;
+import com.qronicle.exception.FileNotFoundException;
 import com.qronicle.exception.ItemNotFoundException;
 import com.qronicle.exception.UserNotFoundException;
 import com.qronicle.model.ItemForm;
-import com.qronicle.service.interfaces.*;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.Authentication;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import javax.validation.constraints.Size;
+import java.time.LocalDate;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
 @Validated
-@RestController
 @CrossOrigin
-@RequestMapping("/items")
+@RestController
+@RequestMapping("${app.api.v1.prefix}/items")
 public class ItemController {
     private final ItemService itemService;
     private final UserService userService;
     private final FileService fileService;
     private final ImageService imageService;
     private final TagService tagService;
+
+    @Value("${app.upload.image.max}")
+    private int MAX_IMAGES;
 
     public ItemController(ItemService itemService, UserService userService, FileService fileService, ImageService imageService, TagService tagService) {
         this.itemService = itemService;
@@ -65,35 +71,24 @@ public class ItemController {
         return ResponseEntity.ok(tags);
     }
 
-    @PreAuthorize("authentication.name == #item.ownerName")
-    @PutMapping
-    public ResponseEntity<Item> updateItem(@RequestBody Item item) {
-        System.out.println(SecurityContextHolder.getContext().getAuthentication().getName());
-        Item oldItem = itemService.findItemById(item.getId());
-        item.setOwner(oldItem.getOwner());
-        item.setTags(oldItem.getTags());
-        itemService.save(item);
-
-        return ResponseEntity.ok(item);
-    }
-
     @PostMapping(consumes = "multipart/form-data")
     public ResponseEntity<Item> addItem(
             @RequestPart("itemForm") ItemForm itemForm,
             @RequestPart(value = "files", required = false)
                 @Size(max = 10, message = "You may upload a maximum of 10 images.")
-                List<MultipartFile> files) {
-        List<Image> images = null;
+                List<MultipartFile> files
+    ) {
+        Set<Image> images = null;
         if (files != null) {
             images =
                 files.stream().map(file -> {
                     String filename = fileService.storeFile(file);
                     return new Image(filename, file.getOriginalFilename(), file.getSize());
-                }).collect(Collectors.toList());
+                }).collect(Collectors.toSet());
         }
-        itemForm.setId(0);
         itemForm.setImages(images);
         itemForm.setUser(userService.getCurrentlyAuthenticatedUser());
+        itemForm.setDate(LocalDate.now());
         Item item = itemService.addItem(itemForm);
         if (item == null && images != null) {
             // failed to create item; delete files in s3 bucket
@@ -102,6 +97,17 @@ public class ItemController {
             }
         }
         return ResponseEntity.status(HttpStatus.CREATED).body(item);
+    }
+
+    //    @PreAuthorize("authentication.name == #item.ownerName")
+    @PutMapping
+    public ResponseEntity<Item> updateItem(@RequestBody Item item) {
+        Item oldItem = itemService.findItemById(item.getId());
+        item.setId(oldItem.getId());
+        item.setOwner(oldItem.getOwner());
+        item.setImages(oldItem.getImages());
+        itemService.save(item);
+        return ResponseEntity.ok(item);
     }
 
     @GetMapping("/user/{username}")
@@ -125,22 +131,49 @@ public class ItemController {
         return ResponseEntity.ok(images);
     }
 
+    // POST method used because CommonsMultipartResolver does not accept PUT method
+    @PostMapping(path = "/{id}/images", consumes = "multipart/form-data")
+    public ResponseEntity<Item> updateItemImages(
+            @RequestPart("item") Item item,
+            @RequestPart("files")
+            @Size(max = 10, message = "You may upload a maximum of 10 images.")
+            List<MultipartFile> files
+    ) {
+        item.setImages(imageService.findImagesByItem(item));
+        if (item.getImages().size() + files.size() > MAX_IMAGES) {
+            throw new RuntimeException("Items cannot have more than " + MAX_IMAGES + " images");
+        }
+        Set<Image> images = new HashSet<>();
+        try {
+            files.forEach(file -> {
+                String filename = fileService.storeFile(file);
+                images.add(new Image(filename, file.getOriginalFilename(), file.getSize()));
+            });
+            images.forEach(item::addImage);
+            itemService.save(item);
+        } catch (Exception e) {
+            System.out.println("ERROR: " + e.getMessage());
+            images.forEach(fileService::deleteFile);
+        }
+
+        return ResponseEntity.ok(item);
+    }
+
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> deleteItem(@PathVariable long id) {
+    public ResponseEntity<?> deleteItem(@PathVariable long id, Authentication auth) {
         Item item = itemService.findItemById(id);
         if (item == null) {
             throw new ItemNotFoundException("No item found with id of " + id);
         }
+        if (!item.getOwner().getUsername().equals(auth.getName())) {
+            throw new AccessDeniedException("You do not have permission to make changes to this item");
+        }
         try {
             Set<Image> images = imageService.findImagesByItem(item);
             // delete all item's images
-            // basic for-loop used to avoid Iterator errors when removing values
-            // TODO: Alter to work with Set
+            images.forEach(System.out::println);
             images.forEach(fileService::deleteFile);
-//            for (int i = 0; i < images.size(); i++) {
-//                Image image = images.stream().;
-//                fileService.deleteFile(image);
-//            }
+            item.setImages(null);
             itemService.delete(item);
         } catch (Exception e) {
             throw new RuntimeException("Error deleting item with id " + id);
@@ -149,21 +182,26 @@ public class ItemController {
         return ResponseEntity.ok("Successfully deleted item with id " + id);
     }
 
-    //@PreAuthorize("authentication.name == #image.item.owner.username")
-    @DeleteMapping("/{itemId}/images")
-    public ResponseEntity<Item> deleteImage(@PathVariable long itemId, @RequestBody Image image) {
+    @DeleteMapping("/{itemId}/images/{imageId}")
+    public ResponseEntity<Item> deleteImage(@PathVariable long itemId, @PathVariable long imageId, Authentication auth) {
         Item item = itemService.findItemById(itemId);
         if (item == null) {
-            throw new ItemNotFoundException("No item found with id of " + itemId);
+            throw new ItemNotFoundException("No item found with ID of " + itemId);
+        }
+        if (!item.getOwner().getUsername().equals(auth.getName())) {
+            throw new AccessDeniedException("You do not have permission to make changes to this item");
+        }
+        Image image = imageService.findImageById(imageId);
+        if (image == null) {
+            throw new FileNotFoundException("No image found with an ID of " + imageId);
         }
          try {
-             image.setItem(null);
              item.removeImage(image);
              fileService.deleteFile(image);
              itemService.save(item);
              return ResponseEntity.ok(item);
          } catch (Exception e) {
-             throw new RuntimeException("Failed to delete the image");
+             throw new RuntimeException("Failed to delete image with ID " + image.getId());
          }
     }
 }
